requirements.txt
asgiref==3.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==5.2.8
gunicorn==22.0.0
idna==3.11
mysqlclient==2.2.7
packaging==25.0
pillow==12.0.0
python-decouple==3.8
qrcode==8.2
requests==2.32.5
sqlparse==0.5.3
typing_extensions==4.15.0
urllib3==2.5.0
